+++ src/model/MessageModel.js
@@ -1,0 +1,33 @@
+const Sequelize = require("sequelize");
+const connection = require("../database/connection");
+
+const Message = connection.define('message', {
+    id: {
+        type: Sequelize.BIGINT,
+        autoIncrement: true,
+        allowNull: false,
+        primaryKey: true
+    },
+    author: {
+        type: Sequelize.TEXT,
+        allowNull: false
+    },
+    content: {
+        type: Sequelize.TEXT,
+        allowNull: false
+    },
+    createdAt: {
+        field: 'created_at',
+        type: Sequelize.DATE,
+        allowNull: false,
+        defaultValue: Sequelize.NOW
+    },
+    updatedAt: {
+        field: 'updated_at',
+        type: Sequelize.DATE,
+        allowNull: false,
+        defaultValue: Sequelize.NOW
+    }
+})
+
+module.exports = Message;